Character 시스템과 ItemSystem 머지 후 잘 작동
11

+++ UC2Team2Project001/Character.cpp
@@ -45,7 +45,7 @@ void Character::TakeDamage(int IncomingDamage)
 		finalDamage = DefenseStrategy->CalculateDamageReceived(this, IncomingDamage);
 	}
 
-	StatManager->ModifyStat(StatType::HP, -finalDamage);
+	StatManager->ModifyStat(StatType::HP, (float) - finalDamage);
 
 	//콜백을 리워드 시스템에 연결
 	auto Event = make_shared<ICharacterDamagedEvent>(CharacterName, finalDamage);
